Moves WalletDbError to a new file under db
