+++ full-service/src/db/mod.rs
@@ -1,7 +1,6 @@
 // Copyright (c) 2020-2021 MobileCoin Inc.
 
 //! Implementations of DB and DB models.
-
 pub mod account;
 pub mod account_txo_status;
 pub mod assigned_subaddress;
@@ -10,38 +9,10 @@ pub mod models;
 pub mod schema;
 pub mod transaction_log;
 pub mod txo;
+mod wallet_db;
+mod wallet_db_error;
 
-use crate::error::WalletDbError;
+
 pub use b58::{b58_decode, b58_encode};
-use diesel::{
-    prelude::*,
-    r2d2::{ConnectionManager, Pool, PooledConnection},
-};
-use mc_common::logger::Logger;
-
-#[derive(Clone)]
-pub struct WalletDb {
-    pool: Pool<ConnectionManager<SqliteConnection>>,
-    logger: Logger,
-}
-
-impl WalletDb {
-    pub fn new(pool: Pool<ConnectionManager<SqliteConnection>>, logger: Logger) -> Self {
-        Self { pool, logger }
-    }
-
-    pub fn new_from_url(database_url: &str, logger: Logger) -> Result<Self, WalletDbError> {
-        let manager = ConnectionManager::<SqliteConnection>::new(database_url);
-        let pool = Pool::builder()
-            .max_size(1)
-            .test_on_check_out(true)
-            .build(manager)?;
-        Ok(Self::new(pool, logger))
-    }
-
-    pub fn get_conn(
-        &self,
-    ) -> Result<PooledConnection<ConnectionManager<SqliteConnection>>, WalletDbError> {
-        Ok(self.pool.get()?)
-    }
-}
+pub use wallet_db::WalletDb;
+pub use wallet_db_error::WalletDbError;
