+++ full-service/src/db/txo.rs
@@ -2,19 +2,16 @@
 
 //! DB impl for the Txo model.
 
-use crate::{
-    db::{
-        account::{AccountID, AccountModel},
-        account_txo_status::AccountTxoStatusModel,
-        assigned_subaddress::AssignedSubaddressModel,
-        b58_encode,
-        models::{
-            Account, AccountTxoStatus, AssignedSubaddress, NewAccountTxoStatus, NewTxo, Txo,
-            TXO_CHANGE, TXO_MINTED, TXO_ORPHANED, TXO_OUTPUT, TXO_PENDING, TXO_RECEIVED,
-            TXO_SECRETED, TXO_SPENT, TXO_UNSPENT,
-        },
+use crate::db::{
+    account::{AccountID, AccountModel},
+    account_txo_status::AccountTxoStatusModel,
+    assigned_subaddress::AssignedSubaddressModel,
+    b58_encode,
+    models::{
+        Account, AccountTxoStatus, AssignedSubaddress, NewAccountTxoStatus, NewTxo, Txo,
+        TXO_CHANGE, TXO_MINTED, TXO_ORPHANED, TXO_OUTPUT, TXO_PENDING, TXO_RECEIVED, TXO_SECRETED,
+        TXO_SPENT, TXO_UNSPENT,
     },
-    error::WalletDbError,
 };
 
 use mc_account_keys::{AccountKey, PublicAddress};
@@ -27,6 +24,7 @@ use mc_transaction_core::{
     tx::{TxOut, TxOutConfirmationNumber},
 };
 
+use crate::db::WalletDbError;
 use diesel::{
     prelude::*,
     r2d2::{ConnectionManager, PooledConnection},
