+++ full-service/src/db/wallet_db.rs
@@ -1,0 +1,33 @@
+use crate::db::WalletDbError;
+use diesel::{
+    prelude::*,
+    r2d2::{ConnectionManager, Pool, PooledConnection},
+};
+use mc_common::logger::Logger;
+
+#[derive(Clone)]
+pub struct WalletDb {
+    pool: Pool<ConnectionManager<SqliteConnection>>,
+    logger: Logger,
+}
+
+impl WalletDb {
+    pub fn new(pool: Pool<ConnectionManager<SqliteConnection>>, logger: Logger) -> Self {
+        Self { pool, logger }
+    }
+
+    pub fn new_from_url(database_url: &str, logger: Logger) -> Result<Self, WalletDbError> {
+        let manager = ConnectionManager::<SqliteConnection>::new(database_url);
+        let pool = Pool::builder()
+            .max_size(1)
+            .test_on_check_out(true)
+            .build(manager)?;
+        Ok(Self::new(pool, logger))
+    }
+
+    pub fn get_conn(
+        &self,
+    ) -> Result<PooledConnection<ConnectionManager<SqliteConnection>>, WalletDbError> {
+        Ok(self.pool.get()?)
+    }
+}
